ci: implement proactive advisory creator
This change introduces a script, advisory-creator.sh, which is used to
generate proactive security advisories for each kernel package updated
in a kernel-kit release.

The script is triggered by new version Git tags pushed to the kernel kit
repository. It then finds all kernel packages updated since the previous
release and generates security advisories for them based on a set of
configuration templates.

The script then opens a pull request, submitting the advisories into
this repository.

Signed-off-by: Sean P. Kelly <seankell@amazon.com>